pyproject.toml
[project]
name = "orthon"
version = "0.1.0"
description = "Diagnostic interpreter for PRISM parquet outputs. Zero calculations."
readme = "README.md"
license = { text = "Proprietary" }
requires-python = ">=3.10"
authors = [
    { name = "Jason", email = "jason@example.com" }
]
keywords = ["diagnostics", "industrial", "parquet", "interpreter"]

dependencies = [
    # Core - PRISM does all the math
    # "prism-engine @ git+https://github.com/prism-engines/prism.git",  # TODO: not yet a Python package

    # Data reading (NOT computation)
    "polars>=0.20.0",
    "duckdb>=0.10.0",
    "pyarrow>=14.0.0",

    # Config
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",

    # Templates/Reports
    "jinja2>=3.1.0",

    # Visualization
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]
ui = [
    "streamlit>=1.30.0",
]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
orthon = "orthon.cli:main"
orthon-explorer = "orthon.explorer.cli:main"
orthon-config = "orthon.data_reader:main"

[project.entry-points."orthon.domains"]
turbofan = "orthon.config.domains:turbofan"
bearings = "orthon.config.domains:bearings"
chemical = "orthon.config.domains:chemical"
hydraulic = "orthon.config.domains:hydraulic"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["orthon"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
